+++ src/parsley/validator.js
@@ -193,6 +193,9 @@ define('parsley/validator', [
           // Again, replace the regexp literal string with the first match group:
           // everything excluding the opening and closing slashes and the flags
           regexp = regexp.replace(new RegExp('^/(.*?)/' + flags + '$'), '$1');
+        } else {
+          // Anchor regexp:
+          regexp = '^' + regexp + '$';
         }
 
         return $.extend(new Validator.Assert().Regexp(regexp, flags), { priority: 64 });